Safari-611.1.10/LayoutTests/fast/events/popup-when-select-change-expected.txt

ALERT: PASSED
 If the pop-up was not blocked then there will be an PASS message. Otherwise, the test fails.
